Successfully change App from jsx back to tsx (fix error caused by 2nd 'null' arg in connect() in List). Render AddCardBtn from App instead of index.tsx.
